# protocol/bc: use pointer receivers