Sometimes function calls can cause allocations when they are on raw
structs instead of pointers. All functions in the bc package have been
changed to use a pointer unless they need to be on raw structs.

```
benchmark                           old ns/op     new ns/op     delta
BenchmarkTxWriteToTrue200-8         260925        217921        -16.48%
BenchmarkTxWriteToFalse200-8        271812        231314        -14.90%
BenchmarkTxInputWriteToTrue-8       822           659           -19.83%
BenchmarkTxInputWriteToFalse-8      901           763           -15.32%
BenchmarkTxOutputWriteToTrue-8      549           486           -11.48%
BenchmarkTxOutputWriteToFalse-8     551           447           -18.87%

benchmark                           old allocs     new allocs   delta
BenchmarkTxWriteToTrue200-8         602            202          -66.45%
BenchmarkTxWriteToFalse200-8        602            202          -66.45%
BenchmarkTxInputWriteToTrue-8       2              1            -50.00%
BenchmarkTxInputWriteToFalse-8      2              1            -50.00%
BenchmarkTxOutputWriteToTrue-8      1              0            -100.00%
BenchmarkTxOutputWriteToFalse-8     1              0            -100.00%

benchmark                           old bytes     new bytes     delta
BenchmarkTxWriteToTrue200-8         32355         320           -99.01%
BenchmarkTxWriteToFalse200-8        32355         327           -98.99%
BenchmarkTxInputWriteToTrue-8       81            1             -98.77%
BenchmarkTxInputWriteToFalse-8      81            1             -98.77%
BenchmarkTxOutputWriteToTrue-8      80            0             -100.00%
BenchmarkTxOutputWriteToFalse-8     80            0             -100.00%
```