@jbowens 2016-11-29
Maybe "protocol/bc: use pointer receivers" for the commit title? That would match the language used here: https://tour.golang.org/methods/4